browseMetadata: Browse and categorise health metadata

Visualise and categorise publicly available metadata from health datasets. By interacting with metadata prior to gaining full access to health datasets, researchers can use this tool to browse datasets and categorise variables.

See also

Author

Maintainer: Rachael Stickland rstickland@turing.ac.uk (ORCID)

Other contributors:

Batool Almarzouq (ORCID) [contributor]

Mahwish Mohammad (ORCID) [contributor]

Daniel Delbarre (ORCID) [contributor]

Nida Ziauddeen (ORCID) [contributor]