compare_sessions

This function is to be used after running the domain_mapping function.

It compares csv outputs from two sessions, finds their differences, and asks for a consensus.

Usage

compare_sessions(
  session_dir,
  session1_base,
  session2_base,
  json_file,
  domain_file
)

Arguments

session_dir

This directory should contain 2 csv files for each session (LOG_ and OUTPUT_), 4 csv files in total.

session1_base

Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-05-16-07-38'

session2_base

Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-08-12-03-30'

json_file

The full path to the metadata file used when running domain_mapping (should be the same for session 1 and session 2)

domain_file

The full path to the domain file used when running domain_mapping (should be the same for session 1 and session 2)

Value

It returns a csv output, which represents the consensus decisions between session 1 and session 2