valid_comparison

Internal Function: This function is called within the map_metadata_compare function.

It reads two inputs to see if they are equal.

If the test is 'warning' status and inputs are not equal it gives warning but continues.

If the test is 'danger' status and inputs are not equal it stops and exits, with error message.

Usage

valid_comparison(input_1, input_2, severity, severity_text)

Arguments

input_1

Input 1

input_2

Input 2

severity

Level of severity. Only 'danger' or 'warning'

severity_text

The text to print if inputs are not equal.

Value

Returns nothing if inputs are equal. If inputs are not equal, returns variable text depending on level of severity.